Docs: align official-support policy with actual pins; extend version test
- changelog #official-support: peer-package wording predated the 0.15.0
  reorg (#2337 dropped peerDependencies); point at config.hugo_version
  and engines.node, note pins track latest, link the minimum.
- Drop obsolete --omit=peer from install commands and rewrite the
  broken Hugo-install tip (no peer deps exist to omit).
- Maintainer notes: rename sections to 'Minimum version' and
  'Officially supported version'; link the official-support policy.
- Rename test to tests/hugo-versions.test.mjs (test:hugo-versions):
  also guards pin-pair sync and minimum <= pin.
