+++ docsy.dev/content/en/blog/2026/0.16.0.md
@@ -151,7 +151,7 @@ themesDir: node_modules
 The install command shape is otherwise unchanged:
 
 ```sh
-npm install --save-dev google/docsy#semver:VERSION --omit=peer
+npm install --save-dev google/docsy#semver:VERSION
 ```
 
 #### Git clone or Git submodule sites {#clone-submodule-sites}
