+++ package.json
@@ -43,7 +43,7 @@
     "check:links": "npm run -C docsy.dev check:links",
     "check:markdown": "npm run _check:markdown -- '**/*.md'",
     "check": "npm run check:format && npm run check:markdown",
-    "ci:post": "npm run fix-for-test && npm run _diff:check && npm run test:hugo-min-sync && npm run test:tooling && npm run test:fixture-site",
+    "ci:post": "npm run fix-for-test && npm run _diff:check && npm run test:hugo-versions && npm run test:tooling && npm run test:fixture-site",
     "ci:prepare": "npm run _prepare && npm run _diff:check",
     "ci:test": "npm run ci:prepare && npm run check && npm run test:website && npm run ci:post",
     "doc-rooted": "TD_BUILD_CTX=doc-rooted npm run",
@@ -68,7 +68,7 @@
     "set:version:git-info": "npm run -s _spv -- --version \"$(scripts/get-build-id.sh)\"",
     "set:version": "npm run -s _spv --",
     "test:fixture-site": "node --test 'tests/fixture-site/**/*.test.mjs'",
-    "test:hugo-min-sync": "node --test tests/hugo-min-sync.test.mjs",
+    "test:hugo-versions": "node --test tests/hugo-versions.test.mjs",
     "test:lychee": "node --test 'tests/lychee/**/*.test.mjs'",
     "test:site-equivalence": "node tests/site-equivalence/full-vs-shared.mjs --check",
     "test:smoke": "node tests/smoke.test.mjs",
