--- tests/hugo-min-sync.test.mjs
@@ -1,52 +1,0 @@
-// The theme's minimum supported Hugo version has three declarations that
-// must agree (see maintainer notes, "Hugo version pins"). Fast and offline.
-
-import test from 'node:test';
-import assert from 'node:assert/strict';
-import fs from 'node:fs';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-const repoRoot = path.resolve(
-  path.dirname(fileURLToPath(import.meta.url)),
-  '..',
-);
-
-const SEMVER = /^\d+\.\d+\.\d+$/;
-
-/** Extract the first regex capture from a file, asserting a match. */
-function extract(relPath, re, what) {
-  const text = fs.readFileSync(path.join(repoRoot, relPath), 'utf8');
-  const m = text.match(re);
-  assert.ok(m, `${what} is declared in ${relPath}`);
-  return m[1];
-}
-
-const declarations = {
-  'theme/theme.toml': () =>
-    extract('theme/theme.toml', /^min_version\s*=\s*"([^"]+)"/m, 'min_version'),
-  'theme/hugo.yaml': () =>
-    extract(
-      'theme/hugo.yaml',
-      /^\s*hugoVersion:\s*\n(?:\s+extended:.*\n)?\s+min:\s*(\S+)/m,
-      'module.hugoVersion.min',
-    ),
-  'docsy.dev/config/_default/hugo.yaml': () =>
-    extract(
-      'docsy.dev/config/_default/hugo.yaml',
-      /^\s*hugoMinVersion:\s*&hugoMinVersion\s+(\S+)/m,
-      'params.hugoMinVersion',
-    ),
-};
-
-test('Hugo minimum-version declarations are in sync', () => {
-  const values = Object.entries(declarations).map(([file, get]) => {
-    const value = get();
-    assert.match(value, SEMVER, `${file} minimum is X.Y.Z semver`);
-    return [file, value];
-  });
-  const [, reference] = values[0];
-  for (const [file, value] of values) {
-    assert.equal(value, reference, `${file} minimum matches theme/theme.toml`);
-  }
-});
